requirements.txt
numpy==1.21.2
pillow==8.3.1
matplotlib==3.4.3